Add files via upload
A line follower robot is a simple yet fascinating type of
mobile robot designed to follow a predefined path, usually
marked by a black or white line on a contrasting surface.
These robots use sensors to detect the line and adjust their
movement accordingly. 

In this project, the ESP32 acts as the brain of the robot,
processing data from sensors and sending commands to
motors to adjust the robot's movement.